requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
websockets==14.2
groq==0.25.0
tavily-python==0.5.0
chromadb==0.6.3
sentence-transformers==4.1.0
pydantic==2.11.1
pydantic-settings==2.8.1
python-dotenv==1.1.0
httpx==0.28.1
